Multiple image upload

+++ app/Http/Controllers/PostsController.php
@@ -3,7 +3,9 @@
 namespace App\Http\Controllers;
 
 use App\Post;
+use App\Image;
 use Auth;
+use GuzzleHttp\Psr7\Response;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Redirect;
 use Illuminate\Support\Facades\Input;
@@ -41,23 +43,42 @@ class PostsController extends Controller
     }
 
     public  function save(Request $request){
+
         $this->validate($request,[
             'title' => 'required',
-            'upload' => 'required',
+            'newfile' => 'required',
             'description'=>'required',
         ]);
-        $image= Input::file('upload');
-        $input['imagename']=time().'.'.$image->getClientOriginalExtension();
-        $destinationPath=public_path('/photos');
-        $image->move($destinationPath,$input['imagename']);
-        $id = Auth::user()->id;
-        $post=new Post([
-            'title'=>$request->get('title'),
-            'description'=>$request->get('description'),
-            'image_upload'=>$input['imagename'],
-            'user_id'=>$id,
-        ]);
-        $post->save();
+            $images = $request->all()["newfile"];
+            $id = Auth::user()->id;
+            $post=new Post([
+                'title'=>$request->get('title'),
+                'description'=>$request->get('description'),
+                'user_id'=>$id,
+            ]);
+            $post->save();
+            foreach ($images as $key=>$image){
+                $input['imagename']=preg_replace('/[^\p{L}\p{N}\s]/u', '', bcrypt(time())).'.'.$image->getClientOriginalExtension();
+//                dump($input['imagename']);
+                $destinationPath=public_path('/photos');
+                $image->move($destinationPath,$input['imagename']);
+                if($image->getClientOriginalName()===$request->get('images')){
+                    $image=new Image([
+                        'image_upload'=>$input['imagename'],
+                        'is_check'=>1,
+                        'post_id'=>$post->id,
+                    ]);
+                    $image->save();
+
+                }else{
+                   $image=new Image([
+                        'image_upload'=>$input['imagename'],
+                        'is_check'=>0,
+                        'post_id'=>$post->id,
+                    ]);
+                    $image->save();
+                }
+            }
         return redirect('posts');
     }
 
@@ -65,14 +86,21 @@ class PostsController extends Controller
     {
         $id=(int)$request->id;
         $post=Post::where('id','=' ,$id)->first();
-        return view('one_post',['post'=>$post]);
+        $images=$post->images;
+//        dd($images);
+        return view('one_post',['post'=>$post,'images'=>$images]);
     }
 
     public function delete(Request $request){
         $id=(int)$request->id;
         $post=Post::where('id','=' ,$id)->first();
-        if($post!=null){
+        $images=$post->images;
+//        dd($images);
+        if($post!=null && $images!=null){
             $post->delete();
+            foreach ($images as $key=>$image){
+                $image->delete();
+            }
         }
         return redirect('posts');
     }
@@ -80,8 +108,10 @@ class PostsController extends Controller
     public function show_hidden_posts(){
         $onlySoftDeleted = Post::onlyTrashed()->get();
 //        foreach ($onlySoftDeleted as $key=>$post){
-//            dd($post->users);
+//            $images[]= $post->trashed_images;
 //        }
+//        dd($images);
+//        foreach ($images as )
 //        dump($onlySoftDeleted);
         return view('soft_posts',['onlySoftDeleted'=>$onlySoftDeleted]);
     }
